+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [1.3.0] - 2024-01-04
 ### Removed
 - Packages `zensols.db.dataclass`, and `zensols.db.dataclass` are no longer
   automatically imported.  You must now import these with their full module
@@ -137,7 +138,8 @@ Functional and downstream moderate risk update release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/dbutil/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/plandes/dbutil/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/plandes/dbutil/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/plandes/dbutil/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/plandes/dbutil/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/plandes/dbutil/compare/v0.2.0...v1.0.0